pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "napi-cli"
version = "0.1.0"
description = "A CLI tool for pulling data from NewsAPI endpoints"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ian Goodrich", email = "ian@igdr.ch"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: News/Diary",
    "Topic :: Utilities",
]
requires-python = ">=3.7"
dependencies = [
    "requests>=2.25.0",
    "python-dotenv>=0.19.0",
    "tqdm>=4.60.0",
]
keywords = ["news", "api", "cli", "newsapi", "eventregistry"]

[project.urls]
Homepage = "https://github.com/specialprocedures/napi"
Repository = "https://github.com/specialprocedures/napi.git"
Issues = "https://github.com/specialprocedures/napi/issues"

[project.scripts]
napi = "napi:cli_main"

[tool.setuptools]
py-modules = ["napi"]